refactor: health utils and legacy health bar
