+++ apps/main/src/llamalend/features/market-position-details/index.ts
@@ -1,8 +1,8 @@
 export * from './BorrowInformation'
 export * from './BorrowPositionDetails'
-export * from './HealthAndBufferBar'
-export * from './HealthBar'
-export * from './HealthDetails'
+export * from './health/HealthAndBufferBar'
+export * from './health/LegacyHealthBar'
+export * from './health/HealthDetails'
 export * from './PositionDetailsComposite'
 export * from './SupplyPositionDetails'
 export * from './MarketEmptyPosition'
